Add basic instrument structure
Instead of sending a boolean to indicate whether the output is muted
send a newly added key event (if key is pressed or released). Add a
trait and struct `Envelope` to control the amplitude based on the
keypress. Add a struct `Instrument` that holds an envelope for
each note that is currently played.