Add proper keyboard layout
Also add a small ascii art of a keyboard to show which keys to press.
There was a bug in the time calculation (probably a rounding error due
to floating points) that caused the frequency to change arbitrarily.
Change the calculation back to use the sample_clock for now. The
absolute time will only become relevant once I'm implementing proper
envelopes.